Refactor to move image conversion details out of cli.py
